Refactor unit operator setup logic to handle RL operator separately - add rl_unit_operator to reinforcement learning part - remove all learning related functions from base unit operator - refactor add_unit function in the world for better readibility and maintanence
